[Server] refactor dir structure + flow fixes
* Better encapsulate server related classes
* Fix a Mongoose static typing regression on callbacks interface
