+++ server/bin/fbTypesGenerator.js
