+++ server/src/server/controllers/ControllerInterface.js
@@ -22,43 +22,16 @@
  * @flow
  */
 
-import type Application from '../services/Application';
+import type Application from '../../services/Application';
 import type {Request, RequestMethod, Response} from 'express';
-import type {Set} from 'immutable';
 
-const AbstractController = require('./AbstractController');
-
-// implement ../ControllerInterface
-class HttpErrorController extends AbstractController {
-
-  code: number;
-  methods: Set<RequestMethod>;
-
-  constructor(application: Application, methods: Set<RequestMethod>, code: number): void {
-    super(application);
-    this.methods = methods;
-    this.code = code;
-  }
-
-  getName(): string {
-    return super.getName() + '-' + this.getCode();
-  }
-
-  getRoute(): string {
-    return '*';
-  }
-
-  getRouteMethods(): Set<RequestMethod> {
-    return this.methods;
-  }
-
-  getCode(): number {
-    return this.code;
-  }
-
-  genResponse(request: Request, response: Response): void {
-    this.returnError(request, response, this.getCode());
-  }
+// eslint-disable-next-line no-undef
+export interface ControllerInterface {
+  constructor(application: Application): void;
+  getName(): string;
+  getApplication(): Application;
+  getRoute(): string;
+  getRouteMethods(): Set<RequestMethod>;
+  onDispatch(request: Request, response: Response): void;
+  genResponse(request: Request, response: Response): void;
 }
-
-module.exports = HttpErrorController;
